Fix ignore list check
The username parameter of is_ignored is being passed a `praw.models.Redditor` not a `str`, so comparing it to the list of strs parsed from the ignore file will always return false. Not sure whether `Redditor.name` will return `/u/aggixx` or just `aggixx` so that might need to be checked to make sure it works. (Or change the ignore list format).